Change courseRate type from BigDecimal to Integer in PurchasedCourses model

+++ backend/src/main/java/org/ezlearn/model/PurchasedCourses.java
@@ -17,7 +17,7 @@ public class PurchasedCourses {
 	private PurchasedCoursesId purchasedCoursesId;
 	
 	private String courseReview;
-	private BigDecimal courseRate;
+	private Integer courseRate;
 	
 
 	public String getCourseReview() {
@@ -26,10 +26,10 @@ public class PurchasedCourses {
 	public void setCourseReview(String courseReview) {
 		this.courseReview = courseReview;
 	}
-	public BigDecimal getCourseRate() {
+	public Integer getCourseRate() {
 		return courseRate;
 	}
-	public void setCourseRate(BigDecimal courseRate) {
+	public void setCourseRate(Integer courseRate) {
 		this.courseRate = courseRate;
 	}
 	public PurchasedCoursesId getPurchasedCoursesId() {
